Update baidu_qianfan_endpoint.py (#94)
* Update modal.py

feat: Raise KeyError when 'prompt' key is missing in JSON response

This commit updates the error handling in the code to raise a KeyError when
the 'prompt' key is not found in the JSON response. This change makes the code
more explicit about the nature of the error, helping to improve clarity and
debugging.

* Update baidu_qianfan_endpoint.py

Import Error was used as Exception, but value error was raised so changed ValueError to ImportError

* Update databricks.py

Updated the raised ValueError to ImportError

---------

Co-authored-by: Aashish Saini <141953346+ShorthillsAI@users.noreply.github.com>